Refactor and fix of instrument diagram:   Now in own folder and split over 11 files instead of 1   Improved routing, both simpler and avoids some bugs   Improved handling of congestion
